requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.1

# Authentication
python-jose[cryptography]==3.3.0
python-jwt==4.0.1
httpx==0.26.0  # For Supabase API calls

# File Processing
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.2.0
pdf2image==1.16.3

# AI Integration
openai==1.12.0
anthropic==0.18.1

# Queue & Workers
celery==5.3.4
redis==5.0.1

# Storage
boto3==1.34.28  # AWS S3
aioboto3==12.2.0  # Async S3

# Payments
stripe==7.8.0

# Reports
reportlab==4.0.9
jinja2==3.1.3

# Email
sendgrid==6.11.0
slowapi==0.1.9  # Alternative rate limiting (optional, using custom for now)

# Utilities
python-dotenv==1.0.1
python-dateutil==2.8.2
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing FastAPI

# Code Quality
black==24.1.1
ruff==0.1.11
mypy==1.8.0

